requirements.txt
google-adk[extensions]>=0.4.0
google-genai>=1.0.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
python-dotenv>=1.0.1
